requirements.txt
#torch==2.4.0+cu118 for py3.10
#torch==2.6.0+cu124 for py3.11
torchvision
torchaudio

accelerate

protobuf
transformers
diffusers
sentencepiece
numpy
scikit-learn
jupyterlab
matplotlib
huggingface_hub
flash_attn
timm
optimum-quantof
git+https://github.com/xhinker/sd_embed.git@main
git+https://github.com/xhinker/azailib.git@main
gradio
modelscope

# required by leffa
fvcore
cloudpickle
av

# required by cosyvoice
HyperPyYAML         #==1.2.2
openai-whisper      #==20231117
inflect             #==7.3.1
WeTextProcessing    #==1.0.3
matcha-tts
pyarrow
torchao

optimum
auto-gptq
bitsandbytes

ftfy
imageio-ffmpeg
imageio

albumentations==0.4.6
webdataset